Fix local packages with no modules being dropped.
If a local package did not have any module, i.e. only an executable with
a single main file, then the package definition would be dropped. This was
due to sloppy code which returned the package by accumulating it along
side each successfully parsed module name. This is unnecessary and silly
and lead to the package definition being discarded if no module names
could be successfully parsed.